Exquisite settings and interactions with spring freshness
Hi, I'm excited to share with you the results of my latest work. In this commit, I've made a number of changes that add new features to your project and increase its efficiency.

Creating a Config package and Beans class is the first step towards a more structured and manageable architecture. I put the Beans class into this package, which is more than just code. It is the key to interacting with Spring, opening up many new perspectives.

I paid special attention to customizing the bean, which now allows your project to work harmoniously with the Spring container. This opens the door to scalability and simplified dependency management.

And in the TestBot class, I revisited the constructor, deciding to remove the service from it. This decision brings more clarity and cleanliness to the code, and emphasizes your bot's ability to interact with the Spring context.

And so, after all these changes, I'm happy to announce that TestBot is now proudly running inside a Spring application, in perfect harmony with its dynamic nature.

I'm proud of this work and hope that these changes will add a new round of development to your project. Let's continue to improve the world of code together!